% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/jtlib.R
\name{readconfig}
\alias{readconfig}
\title{Reading configuration settings from YAML string.}
\usage{
readconfig(configuration)
}
\arguments{
\item{configuration}{string, configuration settings}
}
\value{
list
}
\description{
Reading configuration settings from YAML string.
}

