% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/jtlib.R
\name{writedata}
\alias{writedata}
\title{Writing data to HDF5 file.}
\usage{
writedata(data, data_file)
}
\arguments{
\item{data}{list, data output that should be written to HDF5 file.}

\item{data_file}{string, path to the HDF5 file.}
}
\description{
Writing data to HDF5 file.
}

